kabi.pl: Add support for extra module lists
The extra symbols file is a text file with multiple columns separated by
a single space, the columns are:
- path to module with name
- expected CRC, hex encoded
- name of symbol

Example:
> /path/to/module/lpfc.ko 0xc5850110 printk
> /another/path/to/module/lpfc.ko 0xc5850110 printk

The symbols from the file are checked extra and each mismatching symbol
is reported for each entry in the extra file.

Silencing via rules file has precedence over the extra file.